requirements.txt
Django==3.1
requests==2.24
zwift-client==0.2.0